==> code/0_dataset_statistics.ipynb <==
%config Completer.use_jedi = False


import pandas as pd
from libs.LastNamesInference import LastNamesInference
from tqdm.notebook import tqdm
import matplotlib.pyplot as plt
import numpy as np
import torch

us_papers = pd.read_csv('/data/datasets/WOS/US/US_papers.txt')

df_socsci = pd.read_csv('/data/datasets/WOS/US/text_clean_socsci.txt')
df_health = pd.read_csv('/data/datasets/WOS/US/text_clean_health.txt')


us_papers

Number of articles

us_papers.Annee_Bibliographique.agg({min,max})

len(us_papers.id_art.unique())

Number of Auuthors:

len(us_papers.cluster_ID.unique())

len(df_socsci.id_art.unique())

len(df_health.id_art.unique())
