==> code/1_names_dicts.ipynb <==
import pandas as pd
import numpy as np
import pickle

firstnames = pd.read_excel('../data/names_freq.xlsx',sheet_name='Given names',usecols="A:I")

lastnames = pd.read_excel('../data/names_freq.xlsx',sheet_name='Family names',usecols="A:H")

checks

lastnames.loc[lastnames.name == 'TSETHLIKAI',]

racegrps = ['pctwhite','pctblack','pctapi','pctaian','pct2prace','pcthispanic']

(lastnames[racegrps].sum(axis=1)>1.01).sum()
(lastnames[racegrps].sum(axis=1)<0.9).sum()
lastnames[lastnames[racegrps].sum(axis=1)<0.9]

There are some cases that sum less than 1. This is for privacy preservation. I will normalize so everything sums up to 1

# lastnames[racegrps] = lastnames[racegrps]/lastnames[racegrps].sum(axis=1)
lastnames.loc[:,racegrps] = lastnames.loc[:,racegrps].div(lastnames[racegrps].sum(axis=1), axis=0)

(firstnames[racegrps].sum(axis=1)>101).sum()
(firstnames[racegrps].sum(axis=1)<99).sum()

No problems with first names, but normalize probs to 0-1

firstnames.loc[:,racegrps] = firstnames.loc[:,racegrps].div(firstnames[racegrps].sum(axis=1), axis=0)

Check: Which categories to use

lastnames[['pctwhite','pctblack','pctapi','pctaian','pct2prace','pcthispanic']].multiply(lastnames["count"], axis="index").sum()/lastnames["count"].sum()


I'll keep `'pctwhite','pctblack','pctapi','pcthispanic'` 

# lastnames['pctother'] = lastnames['pctaian'] + lastnames['pct2prace']
# firstnames['pctother'] = firstnames['pctaian'] + firstnames['pct2prace']

remove pctaian and pct2prace and renormalize

lastnames= lastnames.loc[:,['name','count','pctwhite','pcthispanic','pctblack','pctapi']]
lastnames[['pctwhite','pcthispanic','pctblack','pctapi']] = lastnames[['pctwhite','pcthispanic','pctblack','pctapi']].div(
    lastnames[['pctwhite','pcthispanic','pctblack','pctapi']].sum(axis=1), axis=0)

firstnames= firstnames.loc[:,['firstname','obs','pctwhite','pcthispanic','pctblack','pctapi']]
firstnames[['pctwhite','pcthispanic','pctblack','pctapi']] = firstnames[['pctwhite','pcthispanic','pctblack','pctapi']].div(
    firstnames[['pctwhite','pcthispanic','pctblack','pctapi']].sum(axis=1), axis=0)

lastnames  = lastnames.rename({'name':'lastname'},axis=1).set_index('lastname')[['count','pctwhite','pcthispanic','pctblack','pctapi']]
firstnames = firstnames.rename({'obs':'count'},axis=1).set_index('firstname')[['count','pctwhite','pcthispanic','pctblack','pctapi']]

lastnames

#there are some names that only had values for pctaian and pct2prace. I set them to 0
lastnames.loc[lastnames.isna()['pctwhite'].values,]

firstnames.loc[firstnames.isna()['pctwhite'].values,]

lastnames = lastnames.fillna(0)
lastnames.loc[lastnames.isna()['pctwhite'].values,]

lastnames.loc['TSETHLIKAI',]

# First name normalization

#     lastnames_prop = lastnames[['pctwhite','pcthispanic','pctblack','pctapi','pctother']].multiply(lastnames["count"], axis="index").sum()/lastnames["count"].sum()
#     firstnames_prop = firstnames[['pctwhite','pcthispanic','pctblack','pctapi','pctother']].multiply(firstnames["count"], axis="index").sum()/firstnames["count"].sum()
#     ratio = firstnames_general_prop/lastnames_general_prop
#     normalized_names = firstnames[['pctwhite','pcthispanic','pctblack','pctapi','pctother']]/ratio
#     normalized_names = normalized_names.divide(normalized_names.sum(axis=1),axis='index')
#     normalized_names = normalized_names.join(firstnames['count'])

def first_name_normalization(lastnames,firstnames):
    
    lastnames_freq = lastnames[['pctwhite','pcthispanic','pctblack','pctapi']].multiply(lastnames["count"], axis="index")
    firstnames_freq = firstnames[['pctwhite','pcthispanic','pctblack','pctapi']].multiply(firstnames["count"], axis="index")
    
    lastnames_total = lastnames_freq.sum()
    firstnames_total = firstnames_freq.sum()

    expansion_factors = lastnames_total/firstnames_total
    expanded_names = firstnames_freq*expansion_factors
    expanded_names['count'] = expanded_names.sum(axis=1)
    expanded_names[['pctwhite','pcthispanic','pctblack','pctapi']] = expanded_names[['pctwhite','pcthispanic','pctblack','pctapi']].divide(expanded_names['count'], axis= 'index')
    return expanded_names

normalized_names = first_name_normalization(lastnames,firstnames)

lastnames[['pctwhite','pcthispanic','pctblack','pctapi']].multiply(lastnames["count"], axis="index").sum()/lastnames["count"].sum()

firstnames[['pctwhite','pcthispanic','pctblack','pctapi']].multiply(firstnames["count"], axis="index").sum()/firstnames["count"].sum()

normalized_names[['pctwhite','pcthispanic','pctblack','pctapi']].multiply(normalized_names["count"], axis="index").sum()/normalized_names["count"].sum()

normalized_names.to_csv('../data/names_normalized.csv')

## Save

lastnames

lastnames = lastnames.drop('count',axis=1)
firstnames = firstnames.drop('count',axis=1)
normalized_names = normalized_names.drop('count',axis=1)

lastnames_dict = dict(zip(lastnames.index,np.array(lastnames)))
firstnames_dict = dict(zip(firstnames.index,np.array(firstnames)))
firstnames_dict_normalized =  dict(zip(normalized_names.index,np.array(normalized_names)))

dictionaries = {'lastnames_dict':lastnames_dict,'firstnames_dict':firstnames_dict,'firstnames_norm_dict':firstnames_dict_normalized}

I save the dicts

pickle.dump(dictionaries, open('libs/names_prob.p', 'wb'))

Lib/RaceDistribution.py
