==> code/2_authors_race_infer.ipynb <==
import pandas as pd
import re
import random
from tqdm import tqdm
import numpy as np

from libs.RaceDistribution import RaceDistribution
from libs.LastNamesInference import LastNamesInference

us_papers = pd.read_csv('/data/datasets//WOS/US/US_papers.txt')

first_authors = us_papers[us_papers.ordre==1].copy().reset_index(drop=True)
lni = LastNamesInference(names = first_authors.nom)

first_authors.nom.unique().shape

known_names = lni.lastnames_dict.keys()

def clean_prenom(prenom):
    prenom = re.sub(r"\b(?:[A-Z]\.)+(?!\w)",'',prenom)
    prenom = re.sub(' +', ' ',prenom)
    prenom = prenom.strip()
    return prenom.lower()

given_names = [clean_prenom(x) for x in first_authors.Prenom]

all_names = [lni.clean_nom(x) for x in first_authors.nom]

known_names_wos =  [i for i in all_names if i in known_names]
unknown_names_wos =  [i for i in all_names if i not in known_names]

len(unknown_names_wos)

len(unknown_names_wos)/len(all_names)

def infer_race(us_papers):
    first_authors = us_papers[us_papers.ordre==1].copy().reset_index(drop=True)
    #first_authors = first_authors[['id_art','Prenom', 'nom']]
    lni = LastNamesInference(names = first_authors.nom)
    tqdm.pandas(desc="inferring race from lastnames")
    lastname_race_dist = first_authors.progress_apply(lambda x: lni.get_name_dist(lastname=x.nom), axis=1)
    first_authors[lni.prob_order] = pd.DataFrame(lastname_race_dist.to_list())
    first_authors = first_authors[['id_art','white', 'hispanic', 'black', 'asian']]
    us_papers_race = us_papers.merge(first_authors, on ='id_art')

    return us_papers_race

us_papers.sample()

us_papers.Annee_Bibliographique.value_counts()

authors = us_papers[['cluster_ID', 'id_art', 'Prenom', 'nom','ordre']]

first_authors = authors.loc[authors.ordre == 1]

first_authors[['cluster_ID', 'Prenom', 'nom']].nunique()

def clean_prenom(prenom):
    prenom = re.sub(r"\b(?:[A-Z]\.)+(?!\w)",'',prenom)
    prenom = re.sub(' +', ' ',prenom)
    prenom = prenom.strip()
    return prenom.lower()

def clean_nom(nom):
    nom = re.sub(r"-[A-Z]+",'',nom)
    nom = re.sub(' +', ' ',nom)
    nom = nom.strip()
    return nom.lower()

first_authors['Prenom'] = first_authors.Prenom.apply(lambda x: clean_prenom(x))
first_authors['nom'] = first_authors.nom.apply(lambda x: clean_nom(x))

#Select first name
prenom_list = first_authors.groupby(['cluster_ID'])['Prenom'].apply(lambda x: x.values).reset_index()
prenom_list['Prenom'] = prenom_list.Prenom.apply(lambda x: list(filter(None,x)))
prenom_list['Prenom_alt'] = prenom_list.Prenom.apply(lambda x: random.sample(x,1) if len(x)>0 else [''])
prenom_list = prenom_list.explode('Prenom_alt')

first_authors = first_authors.merge(prenom_list[['cluster_ID','Prenom_alt']], how='left', on='cluster_ID')

There's still room for improvement, but I will use this list

#get unique authors list
unique_authors = first_authors.groupby(['cluster_ID', 'Prenom_alt', 'nom']).size().reset_index().rename(columns={0:'count'})

unique_authors.drop('count',axis=1, inplace=True)
unique_authors.columns = ['cluster_ID', 'Prenom','nom']

# I remove duplicates. With the names and lastnames from the cluster this should not be necessary
unique_authors = unique_authors.drop_duplicates('cluster_ID').reset_index(drop=True)

unique_authors = unique_authors.fillna('')

# # exp 1 (sd)
# rd1 = RaceDistribution(exp=1,mode = 'sd')
# authors_dist_1 = unique_authors.progress_apply(lambda x: rd1.get_names_dist(x.Prenom, x.nom), axis=1)
# new_cols = ['exp1_'+ name for name in rd1.prob_order]
# unique_authors[new_cols] = pd.DataFrame(authors_dist_1.to_list())

# exp 2 (var) normalized given names
rd2_norm = RaceDistribution(exp=2,mode = 'sd',normalized_firstnames=True)
authors_dist_2_norm = unique_authors.progress_apply(lambda x: rd2_norm.get_names_dist(x.Prenom, x.nom), axis=1)
new_cols = ['exp2_norm_'+ name for name in rd2_norm.prob_order]
unique_authors[new_cols] = pd.DataFrame(authors_dist_2_norm.to_list())

# exp 2 (var) unnormalized given names
rd2 = RaceDistribution(exp=2,mode = 'sd',normalized_firstnames=False)
authors_dist_2 = unique_authors.progress_apply(lambda x: rd2.get_names_dist(x.Prenom, x.nom), axis=1)
new_cols = ['exp2_notnorm_'+ name for name in rd2_norm.prob_order]
unique_authors[new_cols] = pd.DataFrame(authors_dist_2.to_list())

# Only lastname
rd_lastname = RaceDistribution(mode = 'lastname')
authors_dist_lastname = unique_authors.progress_apply(lambda x: rd_lastname.get_names_dist(x.Prenom, x.nom), axis=1)
new_cols = ['lastname_notnorm_'+ name for name in rd2_norm.prob_order]
unique_authors[new_cols] = pd.DataFrame(authors_dist_lastname.to_list())

# Only first name
rd_name = RaceDistribution(mode = 'name', normalized_firstnames= False)
authors_dist_name = unique_authors.progress_apply(lambda x: rd_name.get_names_dist(x.Prenom, x.nom), axis=1)
new_cols = ['name_notnorm_'+ name for name in rd2_norm.prob_order]
unique_authors[new_cols] = pd.DataFrame(authors_dist_name.to_list())

# Only first name normalized
rd_name_norm = RaceDistribution(mode = 'name', normalized_firstnames=True)
authors_dist_name_norm = unique_authors.progress_apply(lambda x: rd_name_norm.get_names_dist(x.Prenom, x.nom), axis=1)
new_cols = ['name_norm_'+ name for name in rd2_norm.prob_order]
unique_authors[new_cols] = pd.DataFrame(authors_dist_name_norm.to_list())

unique_authors

unique_authors.to_csv('../data/unique_authors.csv', index=False)

# Disciplines

us_papers_race = infer_race(us_papers)

us_papers_race.loc[:,['white','hispanic','black','asian']].mean()

us_papers_race.columns

us_papers_race = us_papers_race[us_papers_race.ordre==1].copy().reset_index(drop=True)

us_papers_race.loc[us_papers_race.gender == 'f','gender'] = 'F'
us_papers_race.loc[us_papers_race.gender == 'm','gender'] = 'M'

us_papers_race = us_papers_race[us_papers_race.gender.isin(['F','M'])].copy().reset_index(drop=True)

# I exclude 'Social Studies of Medicine' because there are too few obs
us_papers_race = us_papers_race[us_papers_race.ESpecialite != 'Social Studies of Medicine']

discipline_gender_agg = us_papers_race.groupby(['EDiscipline','ESpecialite','gender'])['white', 'hispanic', 'black', 'asian'].agg({'count','mean'})

discipline_gender_agg = discipline_gender_agg.stack(level=0)

#discipline_gender_agg.columns = ['EDiscipline', 'ESpecialite', 'gender', 'group', 'count', 'mean']

discipline_gender_agg.index.names = ['EDiscipline', 'ESpecialite', 'gender', 'group']

discipline_gender_agg['freq'] = discipline_gender_agg['count']*discipline_gender_agg['mean']

discipline_gender_agg = discipline_gender_agg.drop(['count','mean'],1)

discipline_gender_agg['joint_prob'] = discipline_gender_agg.freq/ discipline_gender_agg.freq.sum()

discipline_gender_agg

Average citations by group & specialite

df = us_papers_race[['EDiscipline','ESpecialite','gender','cit_rel_all_IAC','cit_all_IAC','white', 'hispanic', 'black', 'asian']].copy()

df = df.dropna(subset = ["cit_all_IAC"], inplace=False).reset_index(drop=True)

df = pd.melt(df, id_vars=['EDiscipline','ESpecialite','gender','cit_all_IAC','cit_rel_all_IAC'], value_vars=['white', 'hispanic', 'black', 'asian'],
       var_name='race', value_name='weights')

def weighted_median(data, weights, quantile=.5):
     # Check the data
    if not isinstance(data, np.matrix):
        data = np.asarray(data)
    if not isinstance(weights, np.matrix):
        weights = np.asarray(weights)
    nd = data.ndim
    if nd != 1:
        raise TypeError("data must be a one dimensional array")
    ndw = weights.ndim
    if ndw != 1:
        raise TypeError("weights must be a one dimensional array")
    if data.shape != weights.shape:
        raise TypeError("the length of data and weights must be the same")
    if ((quantile > 1.) or (quantile < 0.)):
        raise ValueError("quantile must have a value between 0. and 1.")
    # Sort the data
    ind_sorted = np.argsort(data)
    sorted_data = data[ind_sorted]
    sorted_weights = weights[ind_sorted]
    # Compute the auxiliary arrays
    Sn = np.cumsum(sorted_weights)
    # TODO: Check that the weights do not sum zero
    #assert Sn != 0, "The sum of the weights must not be zero"
    Pn = (Sn-0.5*sorted_weights)/Sn[-1]
    # Get the value of the weighted median
    return np.interp(quantile, Pn, sorted_data)

def grouped_wmedian(group):
    data = group['cit_all_IAC']
    weights = group['weights']
    median = weighted_median(data, weights, quantile=.5)
    return median

citations_agg = df.groupby(['EDiscipline','ESpecialite','gender', 'race']).apply(grouped_wmedian)

discipline_gender_agg['median_citations'] = citations_agg

np.all(discipline_gender_agg.median_citations.isnull())

def grouped_wmean(group, citation_type='cit_all_IAC'):
    values = group[citation_type]
    weights = group['weights']
    mean = np.average(values, weights=weights)
    return mean

mean_citations  = df[df.weights!=0].groupby(['EDiscipline','ESpecialite','gender', 'race']).apply(grouped_wmean)

mean_citations

mean_rel_citations  = df[df.weights!=0].groupby(['EDiscipline','ESpecialite','gender', 'race']).apply(lambda x: grouped_wmean(x,'cit_rel_all_IAC'))

mean_rel_citations

discipline_gender_agg['mean_citations'] = mean_citations
discipline_gender_agg['mean_rel_citations'] = mean_rel_citations

discipline_gender_agg.to_csv('../results/discipline_gender_agg.csv')

## Leiden Fields

subfields = pd.read_csv('/data/WOS/US/subfields.txt', delimiter='\t')
titles = pd.read_csv('/data/WOS/US/titles.txt', delimiter='\t')

us_papers_race2 = us_papers_race.merge(titles, how='inner', left_on='id_art', right_on='id_Art')

us_papers_race2 = us_papers_race2.merge(subfields, how='inner', left_on='ITEMID', right_on='ut')

subfields_race = us_papers_race2.groupby('cluster_id1')['white','hispanic','black','asian'].mean().reset_index()

subfields_race.to_csv('../results/leiden_clusters.csv')
