==> code/5_text_clean.ipynb <==
from nltk.corpus import stopwords
from nltk.stem import SnowballStemmer
import string
import pandas as pd
import re
import pickle
from tqdm.notebook import tqdm
import unicodedata2
import numpy as np

from libs.Cleaner import Cleaner


# import nltk
# nltk.download('stopwords')

abstracts = pd.read_csv('/data/WOS/US/abstracts.txt', delimiter='\t')

titles = pd.read_csv('/data/WOS/US/titles.txt', delimiter='\t')

keywords = pd.read_csv('/data/WOS/US/keywords.txt', delimiter='\t')

#normalize names of columns
abstracts.columns = ['id_art', 'Ordre', 'abstract']

titles.columns = ['id_art', 'ITEMID', 'title']

keywords.columns = ['Ident', 'id_art', 'Ordre', 'ItemID', 'keywords']

metadata_df = pd.read_csv('/data/WOS/US/US_papers.txt')

metadata_df[metadata_df.ordre ==1].EDiscipline.value_counts()

# Full DF

# # should change stemming for lemmatizer
# cleaner = Cleaner(abstracts, titles, keywords, stemming=True)

# df_text = cleaner.data_clean()

# df_text

# df = df_text[['id_art','text_clean']]

# df.to_csv('/data/WOS/US/text_clean.txt',index=False)
# df.to_csv('../data/text_clean.txt',index=False)

# df = pd.read_csv('/data/WOS/US/text_clean.txt')

# Social Sciences Dataset

socsci_id = metadata_df.loc[metadata_df.EDiscipline.isin(['Social Sciences','Humanities','Professional Fields']),'id_art']
socsci_id = socsci_id.unique()

abstracts_socsci = abstracts[abstracts.id_art.isin(socsci_id)]
titles_socsci = titles[titles.id_art.isin(socsci_id)]
keywords_socsci = keywords[keywords.id_art.isin(socsci_id)]

cleaner_socsci = Cleaner(abstracts_socsci, titles_socsci, keywords_socsci, stemming=True)

# import nltk
# nltk.download('stopwords')

df_text_socsci = cleaner_socsci.data_clean()

df_socsci = df_text_socsci[['id_art','text_clean']]

df_socsci.to_csv('/data/WOS/US/text_clean_socsci.txt',index=False)
df_socsci.to_csv('../data/text_clean_socsci.txt',index=False)

## Health

health_id = metadata_df.loc[metadata_df.EDiscipline.isin(['Health']),'id_art']
health_id = health_id.unique()

abstracts_health = abstracts[abstracts.id_art.isin(health_id)]
titles_health = titles[titles.id_art.isin(health_id)]
keywords_health = keywords[keywords.id_art.isin(health_id)]

df_text_health = cleaner_health.data_clean()

df_health = df_text_health[['id_art','text_clean']]

df_health

df_health.to_csv('/data/WOS/US/text_clean_health.txt',index=False)
df_health.to_csv('../data/text_clean_health.txt',index=False)
