==> code/8_average_citations.ipynb <==
import pandas as pd
from libs.CitationsProjector import CitationsProjector

import warnings
warnings.filterwarnings("ignore", category=DeprecationWarning)
tqdm.pandas()

## Load data

#df_full = pd.read_csv('/data/WOS/US/text_clean.txt')
df_socsci = pd.read_csv('/data/datasets/WOS/US/text_clean_socsci.txt')
us_papers = pd.read_csv('/data/datasets/WOS/US/US_papers.txt')

df_health = pd.read_csv('/data/datasets/WOS/US/text_clean_health.txt')

df_health.id_art.to_csv('../data/id_art_health.txt',index=False)

# Adding raw citations to the dataset
# raw_citations = pd.read_csv('/data/WOS/US/raw_citations.txt', delimiter='\t', names= ['id_art','cit_all_IAC'])
# us_papers = us_papers.merge(raw_citations,how='left')
# us_papers.to_csv('/data/WOS/US/US_papers.txt', index=False)
# us_papers.to_csv('../data/US_papers.txt', index=False)

#keep only observations with information about citations
us_papers = us_papers.dropna(subset = ["cit_rel_all_IAC","cit_all_IAC"], inplace=False) 

## Social Sciences

cp = CitationsProjector()

average_df, standard_deviation_df, median_df, average_df_norm, standard_deviation_df_norm, median_df_norm = cp.main(us_papers=us_papers,
                                                                                                                    df=df_socsci,
                                                                                                                    dataset_label= 'socsci',
                                                                                                                    n_components=300)

# save results

average_df.to_csv('../results/citations_socsci_mean.csv',index=False)
standard_deviation_df.to_csv('../results/citations_socsci_sd.csv',index=False)
median_df.to_csv('../results/citations_socsci_median.csv',index=False)

## Health

average_df_health,sd_df_health,median_df_health,average_df_norm_health,sd_df_norm_health,median_df_norm_health = cp.main(us_papers=us_papers,
        df=df_health,
        dataset_label= 'health',
        n_components=200)

average_df_health.to_csv('../results/citations_health_mean.csv',index=False)
sd_df_health.to_csv('../results/citations_health_sd.csv',index=False)
median_df_health.to_csv('../results/citations_health_median.csv',index=False)
