==> code/9_topics_robustness.ipynb <==
import string
import pandas as pd
import re
import pickle
from tqdm.notebook import tqdm
import unicodedata2
import numpy as np
import random
import os
from sklearn.metrics.pairwise import cosine_similarity

from libs.lda_wrapper import LDA_wrapper
import logging

import warnings
warnings.filterwarnings("ignore", category=DeprecationWarning)
tqdm.pandas()

df_socsci = pd.read_csv('/data/datasets/WOS/US/text_clean_socsci.txt')

def save(x, file_name):
    with open(file_name, 'wb') as handle:
        pickle.dump(x, handle, protocol=pickle.HIGHEST_PROTOCOL)

def restore(file_name):
    with open(file_name, 'rb') as handle:
        x = pickle.load(handle)
    return x

#random.sample(range(1234, 99999), 10)

def multiple_runs(df,logger,random_states, dataset= 'socsci',n_batches=100,max_iter = 10,n_components=100):
    LDA = LDA_wrapper(n_batches=n_batches)
    
    for rnd in random_states:
        logger.info('running random state {}'.format(rnd))
        texts = df['text_clean'].values
        lda_model,data_vectorized,vectorizer = LDA.lda(data=texts,max_iter = max_iter,n_components=n_components, random_state = rnd)  
        logger.info('finnish running random state {}'.format(rnd))
        save(lda_model, '../results/lda_runs/lda_model_{}_k{}_seed{}.p'.format(dataset,n_components,rnd))
        logger.info('-------------------------------------------------')
    logger.info('finnish running everything')


logging.basicConfig(filename='ldarobusntess_log', filemode="w",format='%(asctime)s %(message)s', datefmt='%d/%m/%Y %H:%M:%S',level=logging.INFO)
logger = logging.getLogger()

random_states = [79613, 5246, 99257, 34310, 68163, 3161, 45052, 10207, 39373, 80243]
#random_states = [45052, 10207, 39373, 80243]

multiple_runs(df_socsci,logger = logger,random_states = random_states,dataset= 'socsci',n_batches=100,max_iter = 10,n_components=300)

restore('../results/lda_runs/lda_model_socsci_k300_seed10207.p')

models_path = os.listdir('../results/lda_runs/')
model_health_path = '../results/lda_model_health_k200.p'

vectorizer_socsci = restore( '../results/vectorizer_socsci.p')
vectorizer_health = restore( '../results/vectorizer_health.p')


def transform_data(df,model_path,vectorizer):    
    texts = df.text_clean.values
    lda_model = restore(model_path)
    data_vectorized = vectorizer.transform(texts)
    doc_dist = lda_model.transform(data_vectorized)
    l2_norm_docs = np.linalg.norm(doc_dist, axis = 1)
    return l2_norm_docs

# Socsci runs
l2_norms = {}

for path in models_path:
    seed = re.search('seed(.*?).p', path).group(1)
    model_path = '../results/lda_runs/' + path

    l2_norm_docs = transform_data(df_socsci, model_path, vectorizer_socsci)

    l2_norms[seed] = l2_norm_docs

#health model
model_path = '../results/lda_model_health_k200.p'

l2_norm_docs = transform_data(df_socsci, model_path, vectorizer_health)
l2_norms['health'] = l2_norm_docs

#random dirichlet ditribution

random_dist = np.random.dirichlet(np.ones(300),size=l2_norms['34310'].shape[0])
ls_norm_random_dist = np.linalg.norm(random_dist, axis = 1)

l2_norms['random'] = ls_norm_random_dist

l2_norms_df = pd.DataFrame.from_dict(l2_norms, orient='index')

#should be 11xN

l2_norms_df

cosine_similarities = cosine_similarity(l2_norms_df)

cosine_similarities_df = pd.DataFrame(cosine_similarities,index = l2_norms_df.index, columns = l2_norms_df.index )

cosine_similarities_df

from sklearn.preprocessing import MinMaxScaler

scaler = MinMaxScaler()

cosine_similarities.min()

min, max = (-1,1)
X_std = (cosine_similarities - cosine_similarities.min()) / (cosine_similarities.max() - cosine_similarities.min())
cosine_similarities_scaled = X_std * (max - min) + min

# np.fill_diagonal(cosine_similarities_scaled, np.nan)

cosine_similarities_df_scaled = pd.DataFrame(cosine_similarities_scaled,index = l2_norms_df.index, columns = l2_norms_df.index )

cosine_similarities_df_scaled

cosine_similarities_df.to_csv('../results/cosine_similarities_multiple_runs.csv')
cosine_similarities_df_scaled.to_csv('../results/cosine_similarities_multiple_runs_scaled.csv')
