==> code/10_lda_all_fields.ipynb <==
import pandas as pd

import numpy as np
from tqdm.notebook import tqdm
import os
import re
import glob
import pickle


from libs.LDADisciplines import DisciplinesCleaner LDA_Fields

from libs.lda_wrapper import LDA_wrapper
from libs.LastNamesInference import LastNamesInference

import warnings
warnings.filterwarnings("ignore", category=DeprecationWarning)
tqdm.pandas()

os.environ["NUMEXPR_NUM_THREADS"] = '127'
os.environ["NUMEXPR_MAX_THREADS"] = '127'

# 1. clean and save text from all disciplines

# dc = DisciplinesCleaner()

# dc.main()

lda_f = LDA_Fields(ntopics=200)

lda_f.fit_all_fields(main_path ="/data/datasets/WOS/US/text/")

# filenames =  glob.glob("/data/datasets/WOS/US/text/*.txt")
# disciplines = [os.path.basename(x).replace(r'text_clean_', '').replace('.txt','') for x in filenames]  
# for discipline,path in tqdm(zip(disciplines,filenames)):
#     tmp = pd.read_csv(path)
#     tmp.sample(1000).to_csv(path.replace('/text/','/text/sample/' ),index=False)
